+++ tests/__init__.py
@@ -5,19 +5,19 @@ from destructify import ParsingContext
 
 
 class DestructifyTestCase(unittest.TestCase):
-    def call_field_to_stream(self, field, value, *, parsed_fields=None):
+    def call_field_to_stream(self, field, value, *, field_values=None):
         stream = io.BytesIO()
-        field.to_stream(stream, value, ParsingContext(parsed_fields=parsed_fields))
+        field.to_stream(stream, value, ParsingContext(field_values=field_values))
         return stream.getvalue()
 
-    def call_field_from_stream(self, field, value, *, parsed_fields=None):
-        return field.from_stream(io.BytesIO(value), ParsingContext(parsed_fields=parsed_fields))
+    def call_field_from_stream(self, field, value, *, field_values=None):
+        return field.from_stream(io.BytesIO(value), ParsingContext(field_values=field_values))
 
     def assertFieldToStreamEqual(self, expected_bytes, python, field, *, parsed_fields=None):
-        self.assertEqual(expected_bytes, self.call_field_to_stream(field, python, parsed_fields=parsed_fields))
+        self.assertEqual(expected_bytes, self.call_field_to_stream(field, python, field_values=parsed_fields))
 
     def assertFieldFromStreamEqual(self, bytes, expected_python, field, *, expected_read=None, parsed_fields=None):
-        res = self.call_field_from_stream(field, bytes, parsed_fields=parsed_fields)
+        res = self.call_field_from_stream(field, bytes, field_values=parsed_fields)
         self.assertEqual(expected_python, res[0])
         if expected_read is not None:
             self.assertEqual(expected_read, res[1])